backend/requirements.txt
# backend/requirements.txt (Docker优化版 v4.0)

# Web框架与服务
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 安全与限流
slowapi>=0.1.9
redis

# 配置与数据模型
pydantic==2.8.2
pydantic-settings
python-dotenv
email-validator

# 外部服务连接
requests
httpx==0.27.0

# 文件处理
python-multipart==0.0.9
aiofiles

# 文档生成与处理
python-docx
# weasyprint  # 注释掉以避免Docker构建问题，需要额外的系统依赖
htmldocx

# LangChain - AI Agent 和工作流
langchain==0.3.15
langchain-openai==0.2.1
langchain-core==0.3.31
langgraph==0.2.45
langgraph-checkpoint==2.0.0
langchain-community==0.3.15

# 向量数据库
chromadb==0.5.0
langchain-chroma==0.1.4

# 文档处理工具
pypandoc
PyPDF2==3.0.1

# PDF 转换功能
reportlab>=3.6.0
# docx2pdf>=0.1.8  # 注释掉以避免Docker构建问题，需要额外的系统依赖

# OCR处理
Pillow
pytesseract

# 文件类型检测
# python-magic  # 注释掉以避免Docker构建问题，需要系统库libmagic1

# --- 搜索工具 ---
duckduckgo-search==6.2.6

# --- 数据库、认证与迁移 ---

# SQLAlchemy ORM 核心库
sqlalchemy

# PostgreSQL 数据库驱动
psycopg2-binary

# PostgreSQL pgvector 扩展 (向量相似度搜索)
pgvector>=0.2.0

# 数据库迁移工具
alembic

# --- 核心修复：明确指定密码库的兼容版本 ---
passlib==1.7.4
bcrypt==4.0.1

# JWT 令牌处理库
python-jose[cryptography]

# --- 开发和测试工具 ---
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
pylint>=2.17.0
bandit[toml]>=1.7.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pyjwt

pdfplumber
rapidocr_pdf
rapidocr_onnxruntime
pdf2docx

# 文件编码检测（用于文本文件处理）
chardet

# PDF 处理和 OCR
pymupdf  # PyMuPDF - 用于 PDF 文字提取和页面渲染

# --- 图表生成（新增） ---
mermaid-py==0.5.0  # Mermaid.js Python 封装
graphviz==0.20.3   # Graphviz Python 封装
# cairosvg==2.7.1  # SVG 转 PNG（可选，需要额外系统依赖）

# --- 任务队列（Celery + Redis） ---
celery==5.3.4
celery-progress==0.5
flower==2.0.1

# --- 监控与指标（Prometheus） ---
prometheus-client>=0.19.0